Campaign::UpdateEraSongUnlockInstructions: residual is MSVC cross-jump placement (2 refuted variants)
99.522 canonical, unchanged.  Drops the redundant (int)(int) double cast
on requiredStars, which is inert (og-dc3-decomp spells it plainly and the
row set is byte-identical before and after).

The whole residual is 6 rows and one fact: four call sites end in an
identical tail

    lwz r4,0(r30) / mr r5,<i8|i9> / mr r3,r22 / bl SetTokenFmt<int> / b exit

(campaign_song_hint_both_singular_move, campaign_song_hint,
campaign_song_hint_both_singular_star, campaign_song_hint_moves).  Both
compilers cross-jump them to ONE physical copy; they disagree only about
WHICH site keeps it.  The image keeps the LAST user's copy at 0x8291DBE4
-- campaign_song_hint_moves falls through into it and the other three
"b" to it.  We keep the FIRST user's copy, at the
campaign_song_hint_both_singular_move site, and the other three branch
back up to it.  Rows 0-289 and 390-429 are identical, so the basic-block
ORDER already matches exactly; only the merged tail's address differs,
and rows [309] [329] [389] are just the three redirected branches.

  [290] tgt b 0x1f10        vs ours bl SetTokenFmt<H>
  [291] tgt --              vs ours b 0x4fe8   (exit)
  [388] tgt bl SetTokenFmt<H> vs ours --
  [309] [329] [389] b 0x1f10/0x1fac vs b 0x4dc8

Refuted this lane (both leave the 6-row set byte-identical):
 1. dropping the (int)(int) cast;
 2. respelling the i8>1 chain's second arm as a plain "if" so it matches
    the i8==1 chain's spelling (semantically identical -- every arm
    returns; i8,i9 are Max(...,0) so the arms are exhaustive).

Checked and clean: name_check shows the same 6 rows and nothing else, so
all 17 function-local-static scope ordinals and every callee name are
already correct (the docs/decomp/patterns/fixable-scope-index.md cost-5
work on this function has landed).  og-dc3-decomp's copy of this function
is character-for-character the same source, so there is no port to take.

Assessment: the cross-jump survivor is chosen from the exit block's
predecessor list, which source order does not reach -- our source order,
block order and register assignment already match the image.  Not
certified at_limit: a lane with a from-scratch rewrite mandate could try
reordering the four arms, but that perturbs the 290-instruction matching
prefix.

+++ src/lazer/meta_ham/Campaign.cpp
@@ -508,7 +508,7 @@ bool Campaign::UpdateEraSongUnlockInstructions(
         return false;
     }
     int requiredStars = pEra->GetSongRequiredStars(song);
-    if ((int)(int)requiredStars == 0) {
+    if (requiredStars == 0) {
         static Symbol era01("era01");
         static Symbol era02("era02");
         static Symbol era03("era03");
